api/requirements.txt
flask
flask-cors
psycopg2
zmq
python-dotenv